refactor: add copy helper & fix copyInput
